requirements.txt
playwright>=1.40.0
python-dotenv>=1.0.0
whatsapp-api-client-python>=0.0.40
aiohttp>=3.9.0
colorlog>=6.7.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
